Ok ∾ prgtest autocommit 13_10 #4

+++ 13_10_estadistiques/Estadistiques.java
@@ -32,7 +32,7 @@ public class Estadistiques {
             valorActual = Integer.parseInt(Entrada.readLine());       
         }
         
-        double mitjana = 0.0;
+        float mitjana = 0;
         
         
         //si la suma de valores és mayor que 0 -- evita dividir por 0
